requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
googledrivedownloader==0.4
idna==3.2
isodate==0.6.0
Jinja2==3.0.1
joblib==1.0.1
littleutils==0.2.2
llvmlite==0.37.0
MarkupSafe==2.0.1
networkx==2.6.3
numba==0.54.0
numpy==1.20.3
ogb==1.3.2
outdated==0.2.1
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
rdflib==6.0.1
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
threadpoolctl==3.0.0
torch==1.9.1+cu111
torch-cluster==1.5.9
torch-geometric==2.0.1
torch-scatter==2.0.8
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
yacs==0.1.8
